+++ src/fr/maxlego08/koth/zcore/utils/ZUtils.java
@@ -10,7 +10,7 @@ import fr.maxlego08.koth.zcore.enums.Permission;
 import fr.maxlego08.koth.zcore.utils.builder.CooldownBuilder;
 import fr.maxlego08.koth.zcore.utils.builder.TimerBuilder;
 import fr.maxlego08.koth.zcore.utils.nms.ItemStackUtils;
-import fr.maxlego08.koth.zcore.utils.nms.NMSUtils;
+import fr.maxlego08.koth.zcore.utils.nms.NmsVersion;
 import fr.maxlego08.koth.zcore.utils.players.ActionBar;
 import net.md_5.bungee.api.chat.BaseComponent;
 import net.md_5.bungee.api.chat.ClickEvent;
@@ -69,12 +69,12 @@ import java.util.stream.Stream;
 @SuppressWarnings("deprecation")
 public abstract class ZUtils extends MessageUtils {
 
-    private static transient List<String> teleportPlayers = new ArrayList<String>();
+    private static final List<String> teleportPlayers = new ArrayList<String>();
     // For plugin support from 1.8 to 1.12
-    private static transient Material[] byId;
+    private static Material[] byId;
 
     static {
-        if (!NMSUtils.isNewVersion()) {
+        if (!NmsVersion.nmsVersion.isNewMaterial()) {
             byId = new Material[0];
             for (Material material : Material.values()) {
                 if (byId.length > material.getId()) {
@@ -129,8 +129,7 @@ public abstract class ZUtils extends MessageUtils {
         PlayerInventory inventory = player.getInventory();
         for (int a = 0; a != 36; a++) {
             ItemStack itemStack = inventory.getContents()[a];
-            if (itemStack == null)
-                slot++;
+            if (itemStack == null) slot++;
         }
         return slot == 0;
     }
@@ -143,10 +142,8 @@ public abstract class ZUtils extends MessageUtils {
      * @param item
      */
     protected void give(Player player, ItemStack item) {
-        if (hasInventoryFull(player))
-            player.getWorld().dropItem(player.getLocation(), item);
-        else
-            player.getInventory().addItem(item);
+        if (hasInventoryFull(player)) player.getWorld().dropItem(player.getLocation(), item);
+        else player.getInventory().addItem(item);
     }
 
     /**
@@ -210,8 +207,7 @@ public abstract class ZUtils extends MessageUtils {
     protected void removeItemInHand(Player player, int how) {
         if (player.getItemInHand().getAmount() > how)
             player.getItemInHand().setAmount(player.getItemInHand().getAmount() - how);
-        else
-            player.setItemInHand(new ItemStack(Material.AIR));
+        else player.setItemInHand(new ItemStack(Material.AIR));
         player.updateInventory();
     }
 
@@ -223,8 +219,7 @@ public abstract class ZUtils extends MessageUtils {
      * @return true if both rentals are the same
      */
     protected boolean same(Location l, Location l2) {
-        return (l.getBlockX() == l2.getBlockX()) && (l.getBlockY() == l2.getBlockY())
-                && (l.getBlockZ() == l2.getBlockZ()) && l.getWorld().getName().equals(l2.getWorld().getName());
+        return (l.getBlockX() == l2.getBlockX()) && (l.getBlockY() == l2.getBlockY()) && (l.getBlockZ() == l2.getBlockZ()) && l.getWorld().getName().equals(l2.getWorld().getName());
     }
 
     /**
@@ -255,15 +250,13 @@ public abstract class ZUtils extends MessageUtils {
         Location playerLocation = player.getLocation();
         AtomicInteger verif = new AtomicInteger(delay);
         teleportPlayers.add(player.getName());
-        if (!location.getChunk().isLoaded())
-            location.getChunk().load();
+        if (!location.getChunk().isLoaded()) location.getChunk().load();
         ses.scheduleWithFixedDelay(() -> {
             if (!same(playerLocation, player.getLocation())) {
                 message(player, Message.TELEPORT_MOVE);
                 ses.shutdown();
                 teleportPlayers.remove(player.getName());
-                if (cmd != null)
-                    cmd.accept(false);
+                if (cmd != null) cmd.accept(false);
                 return;
             }
             int currentSecond = verif.getAndDecrement();
@@ -277,10 +270,8 @@ public abstract class ZUtils extends MessageUtils {
                 teleportPlayers.remove(player.getName());
                 player.teleport(location);
                 message(player, Message.TELEPORT_SUCCESS);
-                if (cmd != null)
-                    cmd.accept(true);
-            } else
-                message(player, Message.TELEPORT_MESSAGE, currentSecond);
+                if (cmd != null) cmd.accept(true);
+            } else message(player, Message.TELEPORT_MESSAGE, currentSecond);
         }, 0, 1, TimeUnit.SECONDS);
     }
 
@@ -343,8 +334,7 @@ public abstract class ZUtils extends MessageUtils {
 
             @Override
             public void run() {
-                if (runnable != null)
-                    runnable.run();
+                if (runnable != null) runnable.run();
             }
         }, delay);
     }
@@ -398,7 +388,7 @@ public abstract class ZUtils extends MessageUtils {
      * @return
      */
     protected double percent(double value, double total) {
-        return (double) ((value * 100) / total);
+        return (value * 100) / total;
     }
 
     /**
@@ -407,7 +397,7 @@ public abstract class ZUtils extends MessageUtils {
      * @return
      */
     protected double percentNum(double total, double percent) {
-        return (double) (total * (percent / 100));
+        return total * (percent / 100);
     }
 
     /**
@@ -465,8 +455,7 @@ public abstract class ZUtils extends MessageUtils {
      * @param page
      * @param objects
      */
-    protected void createInventory(KothPlugin plugin, Player player, EnumInventory inventory, int page,
-                                   Object... objects) {
+    protected void createInventory(KothPlugin plugin, Player player, EnumInventory inventory, int page, Object... objects) {
         plugin.getInventoryManager().createInventory(inventory, player, page, objects);
     }
 
@@ -550,14 +539,13 @@ public abstract class ZUtils extends MessageUtils {
      * @return
      */
     protected String color(String message) {
-        if (message == null)
-            return null;
-        if (NMSUtils.isHexColor()) {
+        if (message == null) return null;
+        if (NmsVersion.nmsVersion.isHexVersion()) {
             Pattern pattern = Pattern.compile("#[a-fA-F0-9]{6}");
             Matcher matcher = pattern.matcher(message);
             while (matcher.find()) {
                 String color = message.substring(matcher.start(), matcher.end());
-                message = message.replace(color, net.md_5.bungee.api.ChatColor.of(color) + "");
+                message = message.replace(color, String.valueOf(net.md_5.bungee.api.ChatColor.of(color)));
                 matcher = pattern.matcher(message);
             }
         }
@@ -571,8 +559,7 @@ public abstract class ZUtils extends MessageUtils {
     protected String colorReverse(String message) {
         if (message == null) return null;
 
-        Pattern pattern = Pattern.compile(net.md_5.bungee.api.ChatColor.COLOR_CHAR + "x[a-fA-F0-9-"
-                + net.md_5.bungee.api.ChatColor.COLOR_CHAR + "]{12}");
+        Pattern pattern = Pattern.compile(net.md_5.bungee.api.ChatColor.COLOR_CHAR + "x[a-fA-F0-9-" + net.md_5.bungee.api.ChatColor.COLOR_CHAR + "]{12}");
         Matcher matcher = pattern.matcher(message);
         while (matcher.find()) {
             String color = message.substring(matcher.start(), matcher.end());
@@ -607,8 +594,7 @@ public abstract class ZUtils extends MessageUtils {
      */
     protected ItemFlag getFlag(String flagString) {
         for (ItemFlag flag : ItemFlag.values()) {
-            if (flag.name().equalsIgnoreCase(flagString))
-                return flag;
+            if (flag.name().equalsIgnoreCase(flagString)) return flag;
         }
         return null;
     }
@@ -681,8 +667,7 @@ public abstract class ZUtils extends MessageUtils {
      * @param command
      * @return
      */
-    protected TextComponent setClickAction(TextComponent component, net.md_5.bungee.api.chat.ClickEvent.Action action,
-                                           String command) {
+    protected TextComponent setClickAction(TextComponent component, net.md_5.bungee.api.chat.ClickEvent.Action action, String command) {
         component.setClickEvent(new ClickEvent(action, command));
         return component;
     }
@@ -692,16 +677,12 @@ public abstract class ZUtils extends MessageUtils {
      * @return
      */
     protected String getDisplayBalence(double value) {
-        if (value < 10000)
-            return format(value, "#.#");
-        else if (value < 1000000)
-            return String.valueOf(Integer.valueOf((int) (value / 1000))) + "k ";
-        else if (value < 1000000000)
-            return String.valueOf(format((value / 1000) / 1000, "#.#")) + "m ";
-        else if (value < 1000000000000l)
-            return String.valueOf(Integer.valueOf((int) (((value / 1000) / 1000) / 1000))) + "M ";
-        else
-            return "to much";
+        if (value < 10000) return format(value, "#.#");
+        else if (value < 1000000) return Integer.valueOf((int) (value / 1000)) + "k ";
+        else if (value < 1000000000) return format((value / 1000) / 1000, "#.#") + "m ";
+        else if (value < 1000000000000L)
+            return Integer.valueOf((int) (((value / 1000) / 1000) / 1000)) + "M ";
+        else return "to much";
     }
 
     /**
@@ -709,16 +690,12 @@ public abstract class ZUtils extends MessageUtils {
      * @return
      */
     protected String getDisplayBalence(long value) {
-        if (value < 10000)
-            return format(value, "#.#");
-        else if (value < 1000000)
-            return String.valueOf(Integer.valueOf((int) (value / 1000))) + "k ";
-        else if (value < 1000000000)
-            return String.valueOf(format((value / 1000) / 1000, "#.#")) + "m ";
-        else if (value < 1000000000000l)
-            return String.valueOf(Integer.valueOf((int) (((value / 1000) / 1000) / 1000))) + "M ";
-        else
-            return "to much";
+        if (value < 10000) return format(value, "#.#");
+        else if (value < 1000000) return Integer.valueOf((int) (value / 1000)) + "k ";
+        else if (value < 1000000000) return format((value / 1000) / 1000, "#.#") + "m ";
+        else if (value < 1000000000000L)
+            return Integer.valueOf((int) (((value / 1000) / 1000) / 1000)) + "M ";
+        else return "to much";
     }
 
     /**
@@ -731,15 +708,13 @@ public abstract class ZUtils extends MessageUtils {
     protected int count(org.bukkit.inventory.Inventory inventory, Material material) {
         int count = 0;
         for (ItemStack itemStack : inventory.getContents())
-            if (itemStack != null && itemStack.getType().equals(material))
-                count += itemStack.getAmount();
+            if (itemStack != null && itemStack.getType().equals(material)) count += itemStack.getAmount();
         return count;
     }
 
     protected Enchantment enchantFromString(String str) {
         for (Enchantment enchantment : Enchantment.values())
-            if (enchantment.getName().equalsIgnoreCase(str))
-                return enchantment;
+            if (enchantment.getName().equalsIgnoreCase(str)) return enchantment;
         return null;
     }
 
@@ -751,8 +726,7 @@ public abstract class ZUtils extends MessageUtils {
 
         direction = direction % 360;
 
-        if (direction < 0)
-            direction += 360;
+        if (direction < 0) direction += 360;
 
         direction = Math.round(direction / 45);
 
@@ -805,8 +779,7 @@ public abstract class ZUtils extends MessageUtils {
      * @return
      */
     protected boolean hasEnchant(Enchantment enchantment, ItemStack itemStack) {
-        return itemStack.hasItemMeta() && itemStack.getItemMeta().hasEnchants()
-                && itemStack.getItemMeta().hasEnchant(enchantment);
+        return itemStack.hasItemMeta() && itemStack.getItemMeta().hasEnchants() && itemStack.getItemMeta().hasEnchant(enchantment);
     }
 
     /**
@@ -835,13 +808,10 @@ public abstract class ZUtils extends MessageUtils {
      */
     protected boolean isCooldown(Player player, String cooldown, int timer) {
         if (CooldownBuilder.isCooldown(cooldown, player)) {
-            ActionBar.sendActionBar(player,
-                    String.format("§cVous devez attendre encore §6%s §cavant de pouvoir faire cette action.",
-                            timerFormat(player, cooldown)));
+            ActionBar.sendActionBar(player, String.format("§cVous devez attendre encore §6%s §cavant de pouvoir faire cette action.", timerFormat(player, cooldown)));
             return true;
         }
-        if (timer > 0)
-            CooldownBuilder.addCooldown(cooldown, player, timer);
+        if (timer > 0) CooldownBuilder.addCooldown(cooldown, player, timer);
         return false;
     }
 
@@ -868,16 +838,12 @@ public abstract class ZUtils extends MessageUtils {
      * @return
      */
     protected String toList(List<String> list, String color, String color2) {
-        if (list == null || list.size() == 0)
-            return null;
-        if (list.size() == 1)
-            return list.get(0);
+        if (list == null || list.size() == 0) return null;
+        if (list.size() == 1) return list.get(0);
         String str = "";
         for (int a = 0; a != list.size(); a++) {
-            if (a == list.size() - 1 && a != 0)
-                str += color + " et " + color2;
-            else if (a != 0)
-                str += color + ", " + color2;
+            if (a == list.size() - 1 && a != 0) str += color + " et " + color2;
+            else if (a != 0) str += color + ", " + color2;
             str += list.get(a);
         }
         return str;
@@ -923,28 +889,22 @@ public abstract class ZUtils extends MessageUtils {
      * @return itemstack
      */
     public ItemStack playerHead(ItemStack itemStack, OfflinePlayer player) {
-        String name = itemStack.hasItemMeta() && itemStack.getItemMeta().hasDisplayName()
-                ? itemStack.getItemMeta().getDisplayName() : null;
-        if (NMSUtils.isNewVersion()) {
+        String name = itemStack.hasItemMeta() && itemStack.getItemMeta().hasDisplayName() ? itemStack.getItemMeta().getDisplayName() : null;
+        if (NmsVersion.nmsVersion.isNewMaterial()) {
             if (itemStack.getType().equals(Material.PLAYER_HEAD) && name != null && name.startsWith("HEAD")) {
                 SkullMeta meta = (SkullMeta) itemStack.getItemMeta();
                 name = name.replace("HEAD", "");
-                if (name.length() == 0)
-                    meta.setDisplayName(null);
-                else
-                    meta.setDisplayName(name);
+                if (name.length() == 0) meta.setDisplayName(null);
+                else meta.setDisplayName(name);
                 meta.setOwningPlayer(player);
                 itemStack.setItemMeta(meta);
             }
         } else {
-            if (itemStack.getType().equals(getMaterial(397)) && itemStack.getData().getData() == 3 && name != null
-                    && name.startsWith("HEAD")) {
+            if (itemStack.getType().equals(getMaterial(397)) && itemStack.getData().getData() == 3 && name != null && name.startsWith("HEAD")) {
                 SkullMeta meta = (SkullMeta) itemStack.getItemMeta();
                 name = name.replace("HEAD", "");
-                if (name.length() == 0)
-                    meta.setDisplayName(null);
-                else
-                    meta.setDisplayName(name);
+                if (name.length() == 0) meta.setDisplayName(null);
+                else meta.setDisplayName(name);
                 meta.setOwner(player.getName());
                 itemStack.setItemMeta(meta);
             }
@@ -958,8 +918,7 @@ public abstract class ZUtils extends MessageUtils {
      * @return itemstack
      */
     protected ItemStack playerHead() {
-        return NMSUtils.isNewVersion() ? new ItemStack(Material.PLAYER_HEAD)
-                : new ItemStack(getMaterial(397), 1, (byte) 3);
+        return NmsVersion.nmsVersion.isNewMaterial() ? new ItemStack(Material.PLAYER_HEAD) : new ItemStack(getMaterial(397), 1, (byte) 3);
     }
 
     /**
@@ -971,9 +930,8 @@ public abstract class ZUtils extends MessageUtils {
      */
     protected <T> T getProvider(Plugin plugin, Class<T> classz) {
         RegisteredServiceProvider<T> provider = plugin.getServer().getServicesManager().getRegistration(classz);
-        if (provider == null)
-            return null;
-        return provider.getProvider() != null ? (T) provider.getProvider() : null;
+        if (provider == null) return null;
+        return provider.getProvider() != null ? provider.getProvider() : null;
     }
 
     /**
@@ -1017,8 +975,7 @@ public abstract class ZUtils extends MessageUtils {
     protected ItemStack createSkull(String url) {
 
         ItemStack head = playerHead();
-        if (url.isEmpty())
-            return head;
+        if (url.isEmpty()) return head;
 
         SkullMeta headMeta = (SkullMeta) head.getItemMeta();
         GameProfile profile = new GameProfile(UUID.randomUUID(), "random_name");
@@ -1045,8 +1002,7 @@ public abstract class ZUtils extends MessageUtils {
      */
     protected boolean isPlayerHead(ItemStack itemStack) {
         Material material = itemStack.getType();
-        if (NMSUtils.isNewVersion())
-            return material.equals(Material.PLAYER_HEAD);
+        if (NmsVersion.nmsVersion.isNewMaterial()) return material.equals(Material.PLAYER_HEAD);
         return (material.equals(getMaterial(397))) && (itemStack.getDurability() == 3);
     }
 
@@ -1059,12 +1015,9 @@ public abstract class ZUtils extends MessageUtils {
      * @throws IllegalArgumentException
      * @throws IllegalAccessException
      */
-    protected Object getPrivateField(Object object, String field)
-            throws SecurityException, NoSuchFieldException, IllegalArgumentException, IllegalAccessException {
+    protected Object getPrivateField(Object object, String field) throws SecurityException, NoSuchFieldException, IllegalArgumentException, IllegalAccessException {
         Class<?> clazz = object.getClass();
-        Field objectField = field.equals("commandMap") ? clazz.getDeclaredField(field)
-                : field.equals("knownCommands") ? NMSUtils.isNewVersion()
-                ? clazz.getSuperclass().getDeclaredField(field) : clazz.getDeclaredField(field) : null;
+        Field objectField = field.equals("commandMap") ? clazz.getDeclaredField(field) : field.equals("knownCommands") ? NmsVersion.nmsVersion.isNewMaterial() ? clazz.getSuperclass().getDeclaredField(field) : clazz.getDeclaredField(field) : null;
         objectField.setAccessible(true);
         Object result = objectField.get(object);
         objectField.setAccessible(false);
@@ -1083,8 +1036,7 @@ public abstract class ZUtils extends MessageUtils {
             SimpleCommandMap commandMap = (SimpleCommandMap) result;
 
             Object map = getPrivateField(commandMap, "knownCommands");
-            @SuppressWarnings("unchecked")
-            HashMap<String, Command> knownCommands = (HashMap<String, Command>) map;
+            @SuppressWarnings("unchecked") HashMap<String, Command> knownCommands = (HashMap<String, Command>) map;
             knownCommands.remove(command.getName());
             for (String alias : command.getAliases()) {
                 knownCommands.remove(alias);
@@ -1106,8 +1058,7 @@ public abstract class ZUtils extends MessageUtils {
     public void glow(ItemStack itemStack) {
         ItemMeta itemMeta = itemStack.getItemMeta();
         itemMeta.addEnchant(Enchantment.ARROW_DAMAGE, 1, true);
-        if (NMSUtils.getNMSVersion() != 1.7)
-            itemMeta.addItemFlags(ItemFlag.HIDE_ENCHANTS);
+        itemMeta.addItemFlags(ItemFlag.HIDE_ENCHANTS);
         itemStack.setItemMeta(itemMeta);
     }
 
@@ -1144,13 +1095,11 @@ public abstract class ZUtils extends MessageUtils {
      * @param notCompletedColor
      * @return string
      */
-    public String getProgressBar(int current, int max, int totalBars, char symbol, String completedColor,
-                                 String notCompletedColor) {
+    public String getProgressBar(int current, int max, int totalBars, char symbol, String completedColor, String notCompletedColor) {
         float percent = (float) current / max;
         int progressBars = (int) (totalBars * percent);
 
-        return color(Strings.repeat(completedColor + symbol, progressBars)
-                + Strings.repeat(notCompletedColor + symbol, totalBars - progressBars));
+        return color(Strings.repeat(completedColor + symbol, progressBars) + Strings.repeat(notCompletedColor + symbol, totalBars - progressBars));
     }
 
     /**
@@ -1162,8 +1111,7 @@ public abstract class ZUtils extends MessageUtils {
      * @return string
      */
     public String getProgressBar(int current, int max, ProgressBar progressBar) {
-        return this.getProgressBar(current, max, progressBar.getLength(), progressBar.getSymbol(),
-                progressBar.getCompletedColor(), progressBar.getNotCompletedColor());
+        return this.getProgressBar(current, max, progressBar.getLength(), progressBar.getSymbol(), progressBar.getCompletedColor(), progressBar.getNotCompletedColor());
     }
 
     /**